Add fpsimd route & refactor
