+++ src/targets/a54x-A546EXXSKFZF4/target.h
@@ -6,7 +6,7 @@
  * ------------------------------------------------------------------------- */
 #define BUILD_VARIANT_LABEL "a54x-A546EXXSKFZF4-app"
 #define PHYS_P0_ORACLE 1
-#define SLIDE_ROUTE SLIDE_ROUTE_MCAST
+#define SLIDE_ROUTE SLIDE_ROUTE_FPSIMD
 
 #ifndef BUILD_FINGERPRINT
 #define BUILD_FINGERPRINT \
@@ -77,7 +77,7 @@
   0x180000ULL, 0x190000ULL, 0x1a0000ULL, 0x1b0000ULL, \
   0x1c0000ULL, 0x1d0000ULL, 0x1e0000ULL, 0x1f0000ULL
 #define SLIDE_MAX_ATTEMPTS 32
-#define SLIDE_ALIGN_MASK 0x7fffULL
+#define SLIDE_KASLR_STEP 0x4000ULL
 #define SLIDE_TRACEFS_EVENT_ID 108
 #define SLIDE_TRACEFS_WORKER_CALLER_OFF 0x0010825cULL
